% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectfkt.R
\name{exerc_fkt}
\alias{exerc_fkt}
\title{Subset variables from data frames}
\usage{
exerc_fkt(A, x)
}
\arguments{
\item{A}{a datatrame}

\item{x}{a vector of either characters (variables names), integers (variables indices), or a     boolean vector}
}
\value{
A data frame with subsetted variables
}
\description{
Subset variables from data frames
}
