% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funktion1.R
\name{my_fancy}
\alias{my_fancy}
\title{Title}
\usage{
my_fancy(x, y)
}
\arguments{
\item{x}{}

\item{y}{}
}
\description{
This will be the description
}
\details{
These will be the details
}
\examples{
my_fancy(2,3)
my_fancy(1:2,3:4)

}
